Split WorkItemsClient into partial classes
